Confirmation of efficient runtime operation compared with old model. Uncertain whether it makes up for the overhead from init, though. Probably not with small input sizes.
Now need to finish documentation and consider adding a simple additional hashing function:

With each iteration, the input can transform a hash. Once it reaches the end of the string, a pointer/reference can be stored at the end. Likelihood of collisions can be reduced through the hashing. In this way, the trie should be able to store true symbol table information for rapid compiler operation.